refactor(repo C5): extract content & document CRUD to engine/repository/content.ts
Moved hashContent, extractTitle, insertContent, insertDocument, upsertDocumentIngestion,
upsertDocumentLinks, getLinkedDocuments, getBacklinkedDocuments, findActiveDocument,
updateDocumentTitle, updateDocument, deactivateDocument, getActiveDocumentPaths,
deleteInactiveDocuments, cleanupOrphanedContent, cleanupOrphanedVectors.
getGraphConnectedCandidates stays in repository.ts because it depends on
parseVirtualPath, which has not yet been extracted.

Build: clean. Tests: 970/970 pass.
