+++ engine/repository/index.ts
@@ -14,3 +14,4 @@ export * from "./chunking.js";
 export * from "./fts.js";
 export * from "./collections.js";
 export * from "./context-annotations.js";
+export * from "./content.js";
